Version fixes

+++ src/opendr/perception/object_detection_2d/centernet/centernet_learner.py
@@ -398,7 +398,7 @@ class CenterNetDetectorLearner(Learner):
             raise ValueError(error_msg)
 
         height, width, _ = _img.shape
-        img_mx = mx.image.image.nd.from_numpy(_img)
+        img_mx = mx.image.image.nd.from_numpy(np.float32(_img))
 
         if keep_size:
             x, img_mx = transform_test(img_mx)
